Around added

+++ app/src/main/java/istanbul/codify/muudy/model/ActivityStat.java
@@ -5,5 +5,5 @@ public final class ActivityStat {
     public int postCount;
     public String post_emoji_word;
     public String post_emoji;
-    public int percent;
+    public double percent;
 }
